requirements.txt
azure-cognitiveservices-vision-face
pillow
opencv-python
firebase_admin